+++ packages/dividends/src/equicast_dividends/__init__.py
@@ -1,8 +1,8 @@
 """equicast-dividends: generic dividend history for any yfinance equity-like symbol."""
 
 from equicast_dividends.client import DividendsClient
-from equicast_dividends.frequency import dividend_frequency
+from equicast_dividends.frequency import dividend_frequency, median_payout_gap_days
 
 __version__ = "0.1.0"
 
-__all__ = ["DividendsClient", "dividend_frequency"]
+__all__ = ["DividendsClient", "dividend_frequency", "median_payout_gap_days"]